requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-ollama>=0.1.0
ollama>=0.1.0
sqlparse>=0.4.0
pytest>=7.0.0
pandas>=1.0.0
allure-pytest>=2.0.0
werkzeug>=2.0.0
configparser>=5.0.0